🏷️ stub `lib._format_impl` private module
